+++ Hub/Project.swift
@@ -84,6 +84,7 @@ let project = Project(
         .local(path: ".."),
         .local(path: "../Shared/SettingsUI"),
         .local(path: "../Shared/HubLink"),
+        .local(path: "../Shared/LaunchChecks"),
         .local(path: "../Applet"),
         .local(path: "../Applet/Protocol"),
         .remote(url: "https://github.com/sparkle-project/Sparkle", requirement: .exact("2.9.4")),
@@ -141,6 +142,7 @@ let project = Project(
             dependencies: [
                 .package(product: "HubCore"),
                 .package(product: "HubLink"),
+                .package(product: "NoodleLaunchChecks"),
                 .package(product: "NoodleRuntimeSettings"),
                 .package(product: "NoodleSettingsUI"),
                 .package(product: "Sparkle"),